+++ src/Components/Cards/RareCard.js
@@ -1,5 +1,5 @@
 
-import capaGen from "../Assets/capaGenerica.png"
+import capaGen from "../../Assets/capaGenerica.png"
 import { CardText, ContainerImg, ImgCard, LiRareCard, RareTicket, TextRare } from "./StylesCards"
 
 const RareCard = ({ title, imgUrl, price }) => {
