--- src/Pages/ShoppingCart/index.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const ShoppingCart = () => {
-  return (
-    <div>index</div>
-  )
-}
-
-export default ShoppingCart
